Move to `ducks` approach in store.
