+++ lib/helpers/dependencies.js
@@ -3,7 +3,7 @@
 import reduce from 'lodash/reduce';
 import find from 'lodash/find';
 
-import { addDependency } from '../store/actions/dependencies';
+import { addDependency } from '../store/dependencies';
 
 import { getEditorKey } from './atom';
