+++ lib/helpers/eventListeners.js
@@ -14,7 +14,7 @@ import {
 
 import { addTooltip, removeTooltips } from '../components/tooltip';
 
-import { selectDependencyByLine } from '../store/selectors/dependencies';
+import { selectDependencyByLine } from '../store/dependencies';
 
 export const onMouseMove = ({ view, atom, store }) => (event) => {
   const regexp = new RegExp(`${NPM_LIBRARY_DESCRIPTION}`, 'g');
