+++ lib/helpers/markers.js
@@ -13,7 +13,7 @@ import {
 
 import { addBadge } from '../components/badge';
 
-import { selectDependencies } from '../store/selectors/dependencies';
+import { selectDependencies } from '../store/dependencies';
 
 export const addMarkerToLine = ({
   atom,
