+++ lib/index.js
@@ -3,7 +3,7 @@
 import { CompositeDisposable, Emitter } from 'atom';
 
 import initStore from './store';
-import subscribeToDependencies from './store/subscriptions/dependencies';
+import { subscribeToDependencies } from './store/dependencies';
 
 import { getEditorKey } from './helpers/atom';
