+++ lib/store/dependencies/actions.js
@@ -5,9 +5,9 @@ import fetch from '../../helpers/fetch';
 import {
   ADD_DEPENDENCY,
   REMOVE_DEPENDENCY,
-} from '../constants/dependencies';
+} from './constants';
 
-import { selectDependency } from '../selectors/dependencies';
+import { selectDependency } from './selectors';
 
 export const addDependency = ({ textEditorKey, line, name }) => (dispatch, getState) => dispatch({
   type: ADD_DEPENDENCY,
