+++ lib/store/dependencies/index.js
@@ -1,0 +1,6 @@
+export * from './constants';
+export * from './actions';
+export * from './selectors';
+
+export { default as subscribeToDependencies } from './subscriptions';
+export { default as dependenciesReducer } from './reducer';
