+++ lib/store/dependencies/reducer.js
@@ -1,6 +1,6 @@
 'use babel';
 
-import { ADD_DEPENDENCY } from '../constants/dependencies';
+import { ADD_DEPENDENCY } from './constants';
 
 const defaultState = {};
