+++ lib/store/dependencies/subscriptions.js
@@ -4,7 +4,7 @@ import isEqual from 'lodash/isEqual';
 import reduce from 'lodash/reduce';
 import forEach from 'lodash/forEach';
 
-import { selectDependency } from '../selectors/dependencies';
+import { selectDependency } from './selectors';
 
 const getUpdatedDependency = (state, nextState) =>
   reduce(nextState, (prev, nextDependency, key) => {
