+++ lib/store/reducers.js
@@ -2,8 +2,8 @@
 
 import { combineReducers } from 'redux';
 
-import dependencies from './dependencies';
+import { dependenciesReducer } from './dependencies';
 
 export default () => combineReducers({
-  dependencies,
+  dependencies: dependenciesReducer,
 });
